Move components ui related classes to the components package
COSDK-500
