+++ core/src/main/java/com/adyen/checkout/core/components/internal/Configuration.kt
@@ -6,7 +6,7 @@
  * Created by ozgur on 2/5/2025.
  */
 
-package com.adyen.checkout.core.internal
+package com.adyen.checkout.core.components.internal
 
 import android.os.Parcelable
 import androidx.annotation.RestrictTo
