+++ core/src/main/java/com/adyen/checkout/core/components/internal/PaymentComponentEvent.kt
@@ -6,7 +6,7 @@
  * Created by ozgur on 16/5/2025.
  */
 
-package com.adyen.checkout.core.internal
+package com.adyen.checkout.core.components.internal
 
 internal sealed class PaymentComponentEvent<ComponentStateT : BaseComponentState> {
