+++ core/src/main/java/com/adyen/checkout/core/components/internal/ui/model/ButtonComponentParams.kt
@@ -6,13 +6,12 @@
  * Created by ozgur on 6/5/2025.
  */
 
-package com.adyen.checkout.core.internal.ui.model
+package com.adyen.checkout.core.components.internal.ui.model
 
 import androidx.annotation.RestrictTo
-import com.adyen.checkout.core.sessions.internal.model.SessionParams
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-data class CommonComponentParamsMapperData(
-    val commonComponentParams: CommonComponentParams,
-    val sessionParams: SessionParams?,
-)
+data class ButtonComponentParams(
+    private val commonComponentParams: CommonComponentParams,
+    override val isSubmitButtonVisible: Boolean,
+) : ComponentParams by commonComponentParams, ButtonParams
