--- core/src/main/java/com/adyen/checkout/core/internal/ButtonConfiguration.kt
@@ -1,16 +1,0 @@
-/*
- * Copyright (c) 2025 Adyen N.V.
- *
- * This file is open source and available under the MIT license. See the LICENSE file for more info.
- *
- * Created by ozgur on 6/5/2025.
- */
-
-package com.adyen.checkout.core.internal
-
-import androidx.annotation.RestrictTo
-
-@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
-interface ButtonConfiguration {
-    val isSubmitButtonVisible: Boolean?
-}
