+++ core/src/main/java/com/adyen/checkout/core/mbway/internal/ui/MBWayDelegateStateFactory.kt
@@ -9,9 +9,9 @@
 package com.adyen.checkout.core.mbway.internal.ui
 
 import com.adyen.checkout.core.common.internal.helper.CountryUtils
-import com.adyen.checkout.core.internal.ui.model.ButtonComponentParams
-import com.adyen.checkout.core.internal.ui.model.ComponentParams
-import com.adyen.checkout.core.internal.ui.model.CountryModel
+import com.adyen.checkout.core.components.internal.ui.model.ButtonComponentParams
+import com.adyen.checkout.core.components.internal.ui.model.ComponentParams
+import com.adyen.checkout.core.components.internal.ui.model.CountryModel
 import com.adyen.checkout.core.internal.ui.state.DelegateStateFactory
 import com.adyen.checkout.core.internal.ui.state.model.DelegateFieldState
 import com.adyen.checkout.core.mbway.internal.ui.model.MBWayDelegateState
