+++ core/src/main/java/com/adyen/checkout/core/mbway/internal/ui/model/MBWayViewState.kt
@@ -8,7 +8,7 @@
 
 package com.adyen.checkout.core.mbway.internal.ui.model
 
-import com.adyen.checkout.core.internal.ui.model.CountryModel
+import com.adyen.checkout.core.components.internal.ui.model.CountryModel
 import com.adyen.checkout.core.internal.ui.state.model.ViewFieldState
 import com.adyen.checkout.core.internal.ui.state.model.toViewFieldState
