+++ core/src/main/java/com/adyen/checkout/core/sessions/CheckoutSessionProvider.kt
@@ -9,7 +9,7 @@
 package com.adyen.checkout.core.sessions
 
 import com.adyen.checkout.core.common.Environment
-import com.adyen.checkout.core.internal.Configuration
+import com.adyen.checkout.core.components.internal.Configuration
 
 internal object CheckoutSessionProvider {
