+++ core/src/main/java/com/adyen/checkout/core/sessions/internal/SessionsComponentEventHandler.kt
@@ -10,9 +10,9 @@ package com.adyen.checkout.core.sessions.internal
 
 import com.adyen.checkout.core.components.CheckoutCallback
 import com.adyen.checkout.core.components.CheckoutResult
+import com.adyen.checkout.core.components.internal.ComponentEventHandler
+import com.adyen.checkout.core.components.internal.PaymentComponentEvent
 import com.adyen.checkout.core.components.paymentmethod.PaymentComponentState
-import com.adyen.checkout.core.internal.ComponentEventHandler
-import com.adyen.checkout.core.internal.PaymentComponentEvent
 import com.adyen.checkout.core.sessions.SessionInteractor
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
