+++ core/src/test/java/com/adyen/checkout/core/mbway/internal/MBWayDelegateStateFactoryTest.kt
@@ -8,8 +8,8 @@
 
 package com.adyen.checkout.core.mbway.internal
 
-import com.adyen.checkout.core.internal.ui.model.ButtonComponentParams
-import com.adyen.checkout.core.internal.ui.model.CountryModel
+import com.adyen.checkout.core.components.internal.ui.model.ButtonComponentParams
+import com.adyen.checkout.core.components.internal.ui.model.CountryModel
 import com.adyen.checkout.core.mbway.internal.ui.MBWayDelegateStateFactory
 import com.adyen.checkout.core.mbway.internal.ui.state.MBWayFieldId
 import org.junit.jupiter.api.Assertions.assertEquals
