+++ core/src/test/java/com/adyen/checkout/core/mbway/internal/model/MBWayStateUpdaterRegistryTest.kt
@@ -8,7 +8,7 @@
 
 package com.adyen.checkout.core.mbway.internal.model
 
-import com.adyen.checkout.core.internal.ui.model.CountryModel
+import com.adyen.checkout.core.components.internal.ui.model.CountryModel
 import com.adyen.checkout.core.internal.ui.state.model.DelegateFieldState
 import com.adyen.checkout.core.mbway.internal.ui.model.MBWayDelegateState
 import com.adyen.checkout.core.mbway.internal.ui.model.MBWayStateUpdaterRegistry
