+++ core/src/test/java/com/adyen/checkout/core/mbway/internal/ui/MBWayComponentStateTest.kt
@@ -10,7 +10,7 @@ package com.adyen.checkout.core.mbway.internal.ui
 
 import com.adyen.checkout.core.components.data.OrderRequest
 import com.adyen.checkout.core.components.data.model.Amount
-import com.adyen.checkout.core.internal.ui.model.CountryModel
+import com.adyen.checkout.core.components.internal.ui.model.CountryModel
 import com.adyen.checkout.core.internal.ui.state.TestFieldTransformerRegistry
 import com.adyen.checkout.core.internal.ui.state.model.DelegateFieldState
 import com.adyen.checkout.core.internal.ui.state.model.Validation
